Add LCOW privileged support
